requirements.txt
absl-py==2.1.0
accelerate==0.32.1
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aioprocessing==2.0.1
aiosignal==1.3.1
altair==5.4.0
analytics-python==1.2.9
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
av==12.3.0
backoff==2.2.1
bigmodelvis==0.0.1
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
deepspeed==0.14.4
distro==1.9.0
docker-pycreds==0.4.0
easydict==1.13
einops==0.6.1
emoji==2.12.1
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.109.2
ffmpy==0.4.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.33.0
google-auth-oauthlib==1.0.0
gql==3.5.0
gradio==3.35.2
gradio_client==0.2.9
graphql-core==3.2.3
grpcio==1.65.4
h11==0.14.0
hjson==3.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.17.3
idna==3.7
imageio==2.34.2
ipdb==0.13.13
ipython==8.26.0
janus==1.0.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.41.1
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.5
narwhals==1.3.0
networkx==3.3
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-ml-py==12.560.30
oauthlib==3.2.2
openai==1.39.0
opencv-python==4.10.0.84
opencv-python-headless==4.9.0.80
orjson==3.10.7
packaging==24.1
pandas==2.2.2
parso==0.8.4
peft==0.3.0
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0
referencing==0.35.1
regex==2024.7.24
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.5.1
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.6.1
sentencepiece==0.1.99
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.7.24
tiktoken==0.7.0
timm==0.6.7
tokenizers==0.14.1
tomli==2.0.1
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
traitlets==5.14.3
transformers==4.35.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.2
uuid_utils==0.9.0
uvicorn==0.27.1
wandb==0.17.7
wcwidth==0.2.13
weave==0.50.14
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
